Introduce aux task pool to the `gtest::ExtManager`
