+++ gtest/src/task_pool.rs
@@ -1,3 +1,23 @@
+// This file is part of Gear.
+
+// Copyright (C) 2024 Gear Technologies Inc.
+// SPDX-License-Identifier: GPL-3.0-or-later WITH Classpath-exception-2.0
+
+// This program is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+
+// This program is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+// GNU General Public License for more details.
+
+// You should have received a copy of the GNU General Public License
+// along with this program. If not, see <https://www.gnu.org/licenses/>.
+
+//! Auxiliary (for tests) task pool implementation for the crate.
+
 use gear_common::{
     auxiliary::{
         taskpool::{AuxiliaryTaskpool, TaskPoolErrorImpl, TaskPoolStorageWrap},
@@ -8,16 +28,14 @@ use gear_common::{
     ProgramId,
 };
 
-/// Task pool manager for gtest environment.
-///
-/// TODO(ap): wait for #4119 and work on integrating this into it, until then
-/// allow(dead_code).
+/// Task pool manager which operates under the hood over [`gear_common::AuxiliaryTaskpool`].
+/// 
+/// Manager is needed mainly to adapt arguments of the task pool methods to the crate.
 #[derive(Debug, Default)]
-#[allow(dead_code)]
 pub(crate) struct TaskPoolManager;
 
-#[allow(dead_code)]
 impl TaskPoolManager {
+    /// Adapted by argument types version of the task pool `add` method.
     pub(crate) fn add(
         &self,
         block_number: BlockNumber,
@@ -26,10 +44,13 @@ impl TaskPoolManager {
         <AuxiliaryTaskpool<TaskPoolCallbacksImpl> as TaskPool>::add(block_number, task)
     }
 
+    /// Adapted by argument types version of the task pool `clear` method.
     pub(crate) fn clear(&self) {
         <AuxiliaryTaskpool<TaskPoolCallbacksImpl> as TaskPool>::clear();
     }
 
+    /// Adapted by argument types version of the task pool `contains` method.
+    #[allow(unused)]
     pub(crate) fn contains(
         &self,
         block_number: &BlockNumber,
@@ -38,6 +59,7 @@ impl TaskPoolManager {
         <AuxiliaryTaskpool<TaskPoolCallbacksImpl> as TaskPool>::contains(block_number, task)
     }
 
+    /// Adapted by argument types version of the task pool `delete` method.
     pub(crate) fn delete(
         &self,
         block_number: BlockNumber,
@@ -46,7 +68,8 @@ impl TaskPoolManager {
         <AuxiliaryTaskpool<TaskPoolCallbacksImpl> as TaskPool>::delete(block_number, task)
     }
 
-    pub(crate) fn drain(
+    /// Adapted by argument types version of the task pool `drain_prefix_keys` method.
+    pub(crate) fn drain_prefix_keys(
         &self,
         block_number: BlockNumber,
     ) -> <TaskPoolStorageWrap as KeyIterableByKeyMap>::DrainIter {
@@ -54,6 +77,7 @@ impl TaskPoolManager {
     }
 }
 
+/// Task pool callbacks implementor.
 pub(crate) struct TaskPoolCallbacksImpl;
 
 impl TaskPoolCallbacks for TaskPoolCallbacksImpl {
@@ -99,14 +123,14 @@ mod tests {
             assert!(manager.contains(&2, task));
         }
 
-        for task in manager.drain(1) {
+        for task in manager.drain_prefix_keys(1) {
             assert!(
                 block_1_tasks.contains(&task),
                 "task not found in block 1 tasks"
             );
         }
 
-        for task in manager.drain(2) {
+        for task in manager.drain_prefix_keys(2) {
             assert!(
                 block_2_tasks.contains(&task),
                 "task not found in block 2 tasks"
